Remove Nalix.Runtime.Pipeline project and refactor code
Removed the Nalix.Runtime.Pipeline project from the solution, including its project file, README, and all solution references. Migrated all relevant source files to new namespaces (Nalix.Runtime.Middleware.*, Nalix.Runtime.Throttling, Nalix.Runtime.Options, Nalix.Runtime.Timekeeping) and updated usings accordingly. This consolidates pipeline, middleware, throttling, and options code under the main runtime, streamlining the codebase and project structure.
