--- src/Nalix.Runtime.Pipeline/Nalix.Runtime.Pipeline.csproj
@@ -1,35 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<Nullable>enable</Nullable>
-		<IsPackable>true</IsPackable>
-		<IsTrimmable>false</IsTrimmable>
-		<OutputType>Library</OutputType>
-		<Deterministic>true</Deterministic>
-		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-		<DisableImplicitMicrosoftCSharpReference>true</DisableImplicitMicrosoftCSharpReference>
-
-		<Description><![CDATA[
-Nalix.Network.Pipeline provides reusable middleware, throttling, and timekeeping components for the Nalix networking stack. It includes inbound packet middleware for permissions, rate limiting, concurrency control, and timeout handling, along with supporting primitives for traffic shaping and clock coordination.
-
-This package is intended for applications and infrastructure that want finer control over request processing pipelines without taking a dependency on the full networking runtime surface.
-		]]></Description>
-
-		<PackageReleaseNotes><![CDATA[
-- Adaptive Traffic Shaping: Robust Token Bucket and Policy-based rate limiters to protect critical network boundaries from traffic bursts.
-- Middleware Defense: High-efficiency inbound filters for permission validation, concurrency gating, and automated timeout enforcement.
-- Temporal Accuracy: Integrated TimeSynchronizer for precise, high-resolution clock coordination in distributed networking environments.
-- Modular Integration: Decoupled pipeline and throttling primitives that can be used independently from the core transport runtime.
-		]]></PackageReleaseNotes>
-
-	</PropertyGroup>
-
-	<Import Project="..\Nalix.Build.props" />
-	<Import Project="..\Package.Build.props" />
-
-	<ItemGroup>
-		<ProjectReference Include="..\Nalix.Common\Nalix.Common.csproj" />
-		<ProjectReference Include="..\Nalix.Framework\Nalix.Framework.csproj" />
-	</ItemGroup>
-</Project>
-
