+++ src/Nalix.Runtime/Middleware/Standard/PermissionMiddleware.cs
@@ -12,10 +12,9 @@ using Nalix.Common.Networking.Protocols;
 using Nalix.Framework.DataFrames.Pooling;
 using Nalix.Framework.DataFrames.SignalFrames;
 using Nalix.Framework.Injection;
-using Nalix.Runtime.Pipeline.Internal;
+using Nalix.Runtime.Middleware.Internal;
 
-
-namespace Nalix.Runtime.Pipeline.Inbound;
+namespace Nalix.Runtime.Middleware.Standard;
 
 /// <summary>
 /// Middleware that checks the permission level of the current connection
