+++ src/Nalix.Runtime/Options/ConcurrencyOptions.cs
@@ -6,7 +6,7 @@ using System.ComponentModel.DataAnnotations;
 using Nalix.Common.Abstractions;
 using Nalix.Framework.Configuration.Binding;
 
-namespace Nalix.Runtime.Pipeline.Options;
+namespace Nalix.Runtime.Options;
 
 /// <summary>
 /// Configuration for the global concurrency gate and circuit breaker.
