+++ src/Nalix.Runtime/Options/DirectiveGuardOptions.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 using Nalix.Common.Abstractions;
 using Nalix.Framework.Configuration.Binding;
 
-namespace Nalix.Runtime.Pipeline.Options;
+namespace Nalix.Runtime.Options;
 
 /// <summary>
 /// Configuration for inbound directive anti-spam guard behavior.
