+++ src/Nalix.Runtime/Options/TokenBucketOptions.cs
@@ -4,7 +4,7 @@
 using Nalix.Common.Abstractions;
 using Nalix.Framework.Configuration.Binding;
 
-namespace Nalix.Runtime.Pipeline.Options;
+namespace Nalix.Runtime.Options;
 
 /// <summary>
 /// Provides configuration options for a high-performance token-bucket rate limiter.
