+++ src/Nalix.Runtime/Throttling/ConcurrencyGate.cs
@@ -18,9 +18,9 @@ using Nalix.Framework.Configuration;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
-using Nalix.Runtime.Pipeline.Options;
+using Nalix.Runtime.Options;
 
-namespace Nalix.Runtime.Pipeline.Throttling;
+namespace Nalix.Runtime.Throttling;
 
 /// <summary>
 /// High-performance per-opcode concurrency limiter with optional FIFO queuing.
