+++ src/Nalix.sln
@@ -17,8 +17,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Analyzers", "Nalix.An
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Analyzers.CodeFixes", "Nalix.Analyzers.CodeFixes\Nalix.Analyzers.CodeFixes.csproj", "{5C2949D2-35AF-4CF5-927E-0FC1B309AE94}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Runtime.Pipeline", "Nalix.Runtime.Pipeline\Nalix.Runtime.Pipeline.csproj", "{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Network.Hosting", "Nalix.Network.Hosting\Nalix.Network.Hosting.csproj", "{95FD7620-8809-4A71-9F79-805C21B925B7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nalix.Runtime", "Nalix.Runtime\Nalix.Runtime.csproj", "{D17F329D-430C-4502-BCB7-F80319CDFB71}"
@@ -117,18 +115,6 @@ Global
 		{5C2949D2-35AF-4CF5-927E-0FC1B309AE94}.Release|x64.Build.0 = Release|Any CPU
 		{5C2949D2-35AF-4CF5-927E-0FC1B309AE94}.Release|x86.ActiveCfg = Release|Any CPU
 		{5C2949D2-35AF-4CF5-927E-0FC1B309AE94}.Release|x86.Build.0 = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|x64.Build.0 = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Debug|x86.Build.0 = Debug|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|x64.ActiveCfg = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|x64.Build.0 = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|x86.ActiveCfg = Release|Any CPU
-		{D5FF83E7-FFFF-4186-AE3D-999C2CB5E62E}.Release|x86.Build.0 = Release|Any CPU
 		{95FD7620-8809-4A71-9F79-805C21B925B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{95FD7620-8809-4A71-9F79-805C21B925B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{95FD7620-8809-4A71-9F79-805C21B925B7}.Debug|x64.ActiveCfg = Debug|Any CPU
